requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.0
requests==2.31.0
PyYAML==6.0.1
pillow==10.2.0
pydantic-settings==2.2.1
